backend finished for iteration 1
