+++ doc/Collabify/1.0.0/controllers_event-controller.js.html
@@ -36,6 +36,7 @@ var helpers 	= require('./helpers');
  * GET /events/:eventId/ - Get current settings for the event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
  * User is the DJ for the requested event &lt;br>
  *
@@ -57,6 +58,7 @@ module.exports.get = function (req, res) {
  * PUT /events/:eventId/ - Change settings for the event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
  * User is the DJ for the requested event &lt;br>
  *
@@ -85,6 +87,7 @@ module.exports.put = function (req, res) {
  * DELETE /events/:eventId/ - End event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
  * User is the DJ for the requested event &lt;br>
  *
@@ -99,25 +102,8 @@ module.exports.put = function (req, res) {
  * @param 				res 					The server response
  */
 module.exports.delete = function (req, res) {
-	helpers.getEventAsDJ(req.headers.userid, req.eventId, res, function (event) {
-		// Remove users from the event
-		User.update({userId: {$in: event.userIds}}, {eventId: null, role: 'NoRole'}, function (err) {
-			if (err) {
-				return status.handleUnexpectedError(err, res);
-			}
-
-			// End the event
-			event.remove();
-
-			// Update the DJ
-			helpers.getUser(req.headers.userid, res, function (user) {
-				user.eventId = null;
-				user.role = 'NoRole';
-				user.save();
-			});
-
-			res.sendStatus(status.OK_DELETE_RESOURCE);
-		});
+	helpers.endEvent(req.headers.userid, req.eventId, res, function () {
+		res.sendStatus(status.OK_DELETE_RESOURCE);
 	});
 };</code></pre>
         </article>
@@ -129,13 +115,13 @@ module.exports.delete = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
