+++ doc/Collabify/1.0.0/controllers_event-user-controller.js.html
@@ -36,6 +36,7 @@ var helpers		= require('./helpers');
  * DELETE /events/:eventId/users/:userId - Leave event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
  * User is at the event &lt;br>
  * User is not the DJ for the event &lt;br>
@@ -45,33 +46,11 @@ var helpers		= require('./helpers');
  * User's eventId is reset to null &lt;br>
  * User's role is reset to 'NoRole' &lt;br>
  *
- * @param 			req 					The client request
- * @param 			req.headers 			The headers in the HTTP request
- * @param {String} 	req.headers.userid 		The user's Spotify ID
- * @param 			res 					The server response
+ * @param 	req		The client request
+ * @param 	res 	The server response
  */
 module.exports.delete = function (req, res) {
-	helpers.getUserAtEvent(req.userId, req.eventId, res, function (user, event) {
-		if (user.role == 'DJ') {
-			logger.error('DJ cannot leave event');
-			return res.sendStatus(status.ERR_UNAUTHORIZED);
-		}
-
-		var userIdIndex = event.userIds.indexOf(user.userId);
-
-		if (userIdIndex == -1) {
-			logger.error('User not at event');
-			return res.sendStatus(status.ERR_BAD_REQUEST);
-		}
-
-		// Remove the user from the event
-		event.userIds.splice(userIdIndex, 1);
-		event.save();
-
-		user.eventId = null;
-		user.role = 'NoRole';
-		user.save();
-
+	helpers.leaveEvent(req.userId, req.eventId, res, function () {
 		res.sendStatus(status.OK_DELETE_RESOURCE);
 	});
 };</code></pre>
@@ -84,13 +63,13 @@ module.exports.delete = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
