+++ doc/Collabify/1.0.0/controllers_event-user-role-controller.js.html
@@ -26,30 +26,54 @@
     
     <section>
         <article>
-            <pre class="prettyprint source linenums"><code>var helpers		= require('./helpers');
+            <pre class="prettyprint source linenums"><code>var logger		= require('../logger');
+var status		= require('../status');
+var helpers		= require('./helpers');
 
 /** @module */
 
 /**
- *
+ * PUT /events/:eventId/users/:userId/role/ - Change user's role
  *
  * &lt;p>Preconditions: &lt;br>
- * User has logged in &lt;br>
+ * Event exists &lt;br>
+ * Requesting user has logged in &lt;br>
+ * Requesting user is the DJ or a Promoted Collabifier &lt;br>
+ * Affected user is at the event &lt;br>
+ * Affected user is not the DJ &lt;br>
+ * New role is not 'DJ' &lt;br>
  *
  * &lt;p>Postconditions: &lt;br>
+ * Affected user has their role changed &lt;br>
  *
- * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
- * @param res The server response
+ * @param 			req 					The client request
+ * @param 			req.headers 			The headers in the HTTP request
+ * @param {String} 	req.headers.userid 		The requestig user's Spotify ID
+ * @param 			req.body 				The body of the request
+ * @param {String} 	req.body.role 			The new role for the affected user
+ * @param 			res 					The server response
  */
 module.exports.put = function (req, res) {
 	// First make sure the DJ is the one making the request
 	helpers.getEventAsDJ(req.headers.userid, req.eventId, res, function () {
 		// Make sure the affected user is at the event
 		helpers.getUserAtEvent(req.userId, req.eventId, res, function (user, event) {
-			user.role = req.body;
+			if (user.role == 'DJ') {
+				logger.error('DJ cannot have their role changed');
+				return res.sendStatus(status.ERR_UNAUTHORIZED);
+			} else if (req.body.role == 'DJ') {
+				logger.error('Cannot promote user to DJ');
+				return res.sendStatus(status.ERR_UNAUTHORIZED);
+			} else if (req.body.role != 'Collabifier'
+					   &amp;&amp; req.body.role != 'Promoted'
+					   &amp;&amp; req.body.role != 'Blacklisted') {
+				logger.error('Invalid role');
+				return res.sendStatus(status.ERR_BAD_REQUEST);
+			}
+
+			user.role = req.body.role;
+			user.save();
+
 			res.sendStatus(status.OK_UPDATE_RESOURCE);
 		});
 	});
@@ -63,13 +87,13 @@ module.exports.put = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
