+++ doc/Collabify/1.0.0/controllers_event-users-controller.js.html
@@ -34,18 +34,22 @@ var helpers		= require('./helpers');
 /** @module */
 
 /**
- *
+ * POST /events/:eventId/users/ - Join event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
+ * User is not already at an event &lt;br>
  *
  * &lt;p>Postconditions: &lt;br>
+ * User is added to the event's user list &lt;br>
+ * User's eventId is changed to match the event &lt;br>
+ * User's role is changed to 'Collabifier' &lt;br>
  *
- * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
- * @param res The server response
+ * @param 			req 					The client request
+ * @param 			req.headers 			The headers in the HTTP request
+ * @param {String} 	req.headers.userid 		The user's Spotify ID
+ * @param 			res 					The server response
  */
 module.exports.post = function (req, res) {
 	helpers.getUser(req.headers.userid, res, function (user) {
@@ -68,18 +72,20 @@ module.exports.post = function (req, res) {
 };
 
 /**
- *
+ * GET /events/:eventId/users/ - Get list of users at event
  *
  * &lt;p>Preconditions: &lt;br>
+ * Event exists &lt;br>
  * User has logged in &lt;br>
+ * User is the DJ for the requested event &lt;br>
  *
- * &lt;p>Postconditions: &lt;br>
- *
- * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
- * @param res The server response
+ * @param 			req 					The client request
+ * @param 			req.headers 			The headers in the HTTP request
+ * @param {String} 	req.headers.userid 		The user's Spotify ID
+ * @param 			res 					The server response
+ * @param {String}	res[i].name				The user's name
+ * @param {String}	res[i].userId			The user's Spotify ID
+ * @param {String}	res[i].role				The user's role
  */
 module.exports.get = function (req, res) {
 	// First, make sure the DJ is the one making the request
@@ -103,13 +109,13 @@ module.exports.get = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
