+++ doc/Collabify/1.0.0/controllers_events-controller.js.html
@@ -43,24 +43,27 @@ module.exports.MAX_EVENT_DISTANCE = 5;
  * POST /events/ - Create a new event
  *
  * &lt;p>Preconditions: &lt;br>
- * User has logged in &lt;br>
  * Event does not already exist &lt;br>
+ * User has logged in &lt;br>
  * User is not at an event already &lt;br>
  *
  * &lt;p>Postconditions: &lt;br>
  * Event is created &lt;br>
  * User is assigned the 'DJ' role for the event &lt;br>
  *
- * @param 				req												The client request
- * @param				req.headers										The headers in the HTTP request
- * @param {String} 		req.headers.userid								The user's Spotify ID
- * @param				req.body										The body of the request
- * @param {String} 		req.body.name									The name of the event
- * @param {Location} 	req.body.location								The event's latitude and longitude
- * @param {String}		[req.body.location.password=null]				The event password (or null if there isn't one)
- * @param {Boolean}		[req.body.location.locationRestricted=true]		Whether to restrict the event to nearby users
- * @param {Boolean}		[req.body.location.allowVoting=true]			Whether to allow users to vote on songs
- * @param 				res												The server response
+ * @param 						req												The client request
+ * @param						req.headers										The headers in the HTTP request
+ * @param {String} 				req.headers.userid								The user's Spotify ID
+ * @param						req.body										The body of the request
+ * @param {String} 				req.body.name									The name of the event
+ * @param {Location}			req.body.location								The location of the event
+ * @param {Number}				req.body.location.latitude						The latitude of the event
+ * @param {Number}				req.body.location.longitude						The longitude of the event
+ * @param {EventSettings}		req.body.settings								The settings for the event
+ * @param {String}				[req.body.settings.password=null]				The event password (or null if there isn't one)
+ * @param {Boolean}				[req.body.settings.locationRestricted=true]		Whether to restrict the event to nearby users
+ * @param {Boolean}				[req.body.settings.allowVoting=true]			Whether to allow users to vote on songs
+ * @param 						res												The server response
  */
 module.exports.post = function (req, res) {
 	helpers.getUser(req.headers.userid, res, function (user) {
@@ -95,14 +98,22 @@ module.exports.post = function (req, res) {
 /**
  * GET /events/ - Get all events near provided coordinates
  *
- * @param 				req 						The client request
- * @param				req.headers					The headers in the HTTP request
- * @param {String}		req.headers.latitude		The user's current latitude
- * @param {String} 		req.headers.longitude		The user's current longitude
- * @param {Event[]} 	res 						The server response - list of all nearby events
+ * @param 						req 									The client request
+ * @param						req.headers								The headers in the HTTP request
+ * @param {String}				req.headers.latitude					The user's current latitude
+ * @param {String} 				req.headers.longitude					The user's current longitude
+ * @param 						res 									The server response
+ * @param {String}				res[i].name								The event's name
+ * @param {String}				res[i].eventId							The Spotify ID for the user who created the event
+ * @param {Location}			res[i].location							The location of the event
+ * @param {String}				res[i].location.latitude				The latitude of the event
+ * @param {String}				res[i].location.longitude				The longitude of the event
+ * @param {EventSettings}		res[i].settings							The settings for the event
+ * @param {String}				res[i].settings.password				The event password (or null if there isn't one)
+ * @param {Boolean}				res[i].settings.locationRestricted		Whether to restrict the event to nearby users
+ * @param {Boolean}				res[i].settings.allowVoting				Whether to allow users to vote on songs
  */
 module.exports.get = function (req, res) {
-	/** @todo Verify latitude and longitude get converted properly */
 	var latitude = Number(req.headers.latitude);
 	var longitude = Number(req.headers.longitude);
 
@@ -133,13 +144,13 @@ module.exports.get = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
