+++ doc/Collabify/1.0.0/controllers_user-controller.js.html
@@ -33,18 +33,19 @@ var helpers		= require('./helpers');
 /** @module */
 
 /**
- *
+ * GET /users/:userId/ - Get user details
  *
  * &lt;p>Preconditions: &lt;br>
  * User has logged in &lt;br>
  *
- * &lt;p>Postconditions: &lt;br>
- *
- * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
- * @param res The server response
+ * @param 					req 						The client request
+ * @param 					res 						The server response
+ * @param {String} 			res.name 					The user's name
+ * @param {String} 			res.userId 					The user's Spotify ID
+ * @param {String} 			res.eventId 				The ID of the event the user is at, or null if they aren't at one
+ * @param {String} 			res.role 					The user's role
+ * @param {UserSettings} 	res.settings 				The user's current settings
+ * @param {Boolean} 		res.settings.showName		Whether to display the user's Spotify username or 'anonymous'
  */
 module.exports.get = function (req, res) {
 	helpers.getUser(req.userId, 'name userId eventId role settings', res, function (user) {
@@ -53,18 +54,18 @@ module.exports.get = function (req, res) {
 };
 
 /**
- *
+ * PUT /users/:userId/ - Change user settings
  *
  * &lt;p>Preconditions: &lt;br>
  * User has logged in &lt;br>
  *
  * &lt;p>Postconditions: &lt;br>
+ * User's settings are updated
  *
- * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
- * @param res The server response
+ * @param 				req 					The client request
+ * @param 				req.body 				The body of the request
+ * @param {String} 		req.body.showName 		Whether to display the user's Spotify username or 'anonymous'
+ * @param 				res 					The server response
  */
 module.exports.put = function (req, res) {
 	helpers.getUser(req.userId, res, function (user) {
@@ -74,42 +75,35 @@ module.exports.put = function (req, res) {
 };
 
 /**
- *
+ * DELETE /users/:userId/ - Log out the user and delete them from the database
  *
  * &lt;p>Preconditions: &lt;br>
  * User has logged in &lt;br>
  *
  * &lt;p>Postconditions: &lt;br>
+ * User is removed from the database &lt;br>
+ * If the user was a DJ for an event, the event is ended &lt;br>
+ * If the user was at an event, they are removed from the event's user list &lt;br>
  *
  * @param req The client request
- * @param req.headers The headers in the HTTP request
- * @param {String} req.headers.userid The user's Spotify ID
- * @param req.body The body of the request
  * @param res The server response
  */
 module.exports.delete = function (req, res) {
 	helpers.getUser(req.userId, res, function (user) {
-		// Before logging out the user, remove them from their event if they were
-		// at one
-		if (user.eventId != null) {
-			helpers.getEvent(user.eventId, res, function (event) {
-				var userIdIndex = event.userIds.indexOf(user.userId);
-
-				if (userIdIndex == -1) {
-					logger.error('User not at event');
-					return res.sendStatus(status.ERR_BAD_REQUEST);
-				}
-
-				// Remove the user from the event
-				event.userIds.splice(userIdIndex, 1);
-				event.save();
-			});
+		var logOut = function () {
+			user.remove();
+			res.sendStatus(status.OK_DELETE_RESOURCE);
 		}
 
-		// Log out the user
-		user.remove();
+		if (user.eventId != null) {
+			if (user.role == 'DJ') {
+				return helpers.endEvent(user.userId, user.eventId, res, logOut);
+			} else {
+				return helpers.leaveEvent(user.userId, user.eventId, res, logOut);
+			}
+		}
 
-		res.sendStatus(status.OK_DELETE_RESOURCE);
+		logOut();
 	});
 };</code></pre>
         </article>
@@ -121,13 +115,13 @@ module.exports.delete = function (req, res) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
