+++ doc/Collabify/1.0.0/module-controllers_event-controller.html
@@ -71,6 +71,7 @@
     DELETE /events/:eventId/ - End event
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is the DJ for the requested event <br>
 
@@ -276,7 +277,7 @@ Users at the event (including the DJ) have their role reset to 'NoRole' <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line73">line 73</a>
+        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line76">line 76</a>
     </li></ul></dd>
     
 
@@ -317,6 +318,7 @@ Users at the event (including the DJ) have their role reset to 'NoRole' <br>
     GET /events/:eventId/ - Get current settings for the event
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is the DJ for the requested event <br>
 </div>
@@ -612,7 +614,7 @@ User is the DJ for the requested event <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line22">line 22</a>
+        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line23">line 23</a>
     </li></ul></dd>
     
 
@@ -653,6 +655,7 @@ User is the DJ for the requested event <br>
     PUT /events/:eventId/ - Change settings for the event
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is the DJ for the requested event <br>
 
@@ -969,7 +972,7 @@ The event's settings are updated <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line47">line 47</a>
+        <a href="controllers_event-controller.js.html">controllers/event-controller.js</a>, <a href="controllers_event-controller.js.html#line49">line 49</a>
     </li></ul></dd>
     
 
@@ -1011,13 +1014,13 @@ The event's settings are updated <br>
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
