+++ doc/Collabify/1.0.0/module-controllers_event-playlist-controller.html
@@ -71,6 +71,7 @@
     GET /events/:eventId/playlist/ - Get all songs in the playlist in their proper order
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is at the event <br>
 </div>
@@ -455,7 +456,7 @@ User is at the event <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line77">line 77</a>
+        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line79">line 79</a>
     </li></ul></dd>
     
 
@@ -496,6 +497,7 @@ User is at the event <br>
     POST /events/:eventId/playlist/ - Add song to playlist
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is at the event <br>
 User is not blacklisted <br>
@@ -860,7 +862,7 @@ Song is added to the end of the playlist <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line30">line 30</a>
+        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line31">line 31</a>
     </li></ul></dd>
     
 
@@ -901,6 +903,7 @@ Song is added to the end of the playlist <br>
     PUT /events/:eventId/playlist/ - Reorder songs in the playlist
 
 <p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
 User is the DJ or a Promoted Collabifier for the requested event <br>
 
@@ -1434,7 +1437,7 @@ Songs in the playlist are reordered accordingly
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line107">line 107</a>
+        <a href="controllers_event-playlist-controller.js.html">controllers/event-playlist-controller.js</a>, <a href="controllers_event-playlist-controller.js.html#line110">line 110</a>
     </li></ul></dd>
     
 
@@ -1476,13 +1479,13 @@ Songs in the playlist are reordered accordingly
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
