+++ doc/Collabify/1.0.0/module-controllers_event-users-controller.html
@@ -68,10 +68,12 @@
 
 
 <div class="description">
-    <p>Preconditions: <br>
-User has logged in <br>
+    GET /events/:eventId/users/ - Get list of users at event
 
-<p>Postconditions: <br>
+<p>Preconditions: <br>
+Event exists <br>
+User has logged in <br>
+User is the DJ for the requested event <br>
 </div>
 
 
@@ -208,24 +210,6 @@ User has logged in <br>
         </tr>
 
     
-
-        <tr>
-            
-                <td class="name"><code>body</code></td>
-            
-
-            <td class="type">
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The body of the request</td>
-        </tr>
-
-    
     </tbody>
 </table>
 
@@ -248,6 +232,75 @@ User has logged in <br>
             
 
             <td class="description last">The server response</td>
+        </tr>
+
+    
+
+        <tr>
+            
+                <td class="name"><code>res[i].name</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">String</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">The user's name</td>
+        </tr>
+
+    
+
+        <tr>
+            
+                <td class="name"><code>res[i].userId</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">String</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">The user's Spotify ID</td>
+        </tr>
+
+    
+
+        <tr>
+            
+                <td class="name"><code>res[i].role</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">String</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">The user's role</td>
         </tr>
 
     
@@ -288,7 +341,7 @@ User has logged in <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-users-controller.js.html">controllers/event-users-controller.js</a>, <a href="controllers_event-users-controller.js.html#line56">line 56</a>
+        <a href="controllers_event-users-controller.js.html">controllers/event-users-controller.js</a>, <a href="controllers_event-users-controller.js.html#line62">line 62</a>
     </li></ul></dd>
     
 
@@ -326,10 +379,17 @@ User has logged in <br>
 
 
 <div class="description">
-    <p>Preconditions: <br>
+    POST /events/:eventId/users/ - Join event
+
+<p>Preconditions: <br>
+Event exists <br>
 User has logged in <br>
+User is not already at an event <br>
 
 <p>Postconditions: <br>
+User is added to the event's user list <br>
+User's eventId is changed to match the event <br>
+User's role is changed to 'Collabifier' <br>
 </div>
 
 
@@ -466,24 +526,6 @@ User has logged in <br>
         </tr>
 
     
-
-        <tr>
-            
-                <td class="name"><code>body</code></td>
-            
-
-            <td class="type">
-            
-            </td>
-
-            
-
-            
-
-            <td class="description last">The body of the request</td>
-        </tr>
-
-    
     </tbody>
 </table>
 
@@ -546,7 +588,7 @@ User has logged in <br>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="controllers_event-users-controller.js.html">controllers/event-users-controller.js</a>, <a href="controllers_event-users-controller.js.html#line22">line 22</a>
+        <a href="controllers_event-users-controller.js.html">controllers/event-users-controller.js</a>, <a href="controllers_event-users-controller.js.html#line26">line 26</a>
     </li></ul></dd>
     
 
@@ -588,13 +630,13 @@ User has logged in <br>
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-controllers_event-controller.html">controllers/event-controller</a></li><li><a href="module-controllers_event-playlist-controller.html">controllers/event-playlist-controller</a></li><li><a href="module-controllers_event-playlist-song-controller.html">controllers/event-playlist-song-controller</a></li><li><a href="module-controllers_event-user-controller.html">controllers/event-user-controller</a></li><li><a href="module-controllers_event-user-role-controller.html">controllers/event-user-role-controller</a></li><li><a href="module-controllers_event-users-controller.html">controllers/event-users-controller</a></li><li><a href="module-controllers_events-controller.html">controllers/events-controller</a></li><li><a href="module-controllers_helpers.html">controllers/helpers</a></li><li><a href="module-controllers_user-controller.html">controllers/user-controller</a></li><li><a href="module-controllers_users-controller.html">controllers/users-controller</a></li><li><a href="module-models_event.html">models/event</a></li><li><a href="module-models_event-settings.html">models/event-settings</a></li><li><a href="module-models_location.html">models/location</a></li><li><a href="module-models_playlist.html">models/playlist</a></li><li><a href="module-models_song.html">models/song</a></li><li><a href="module-models_user.html">models/user</a></li><li><a href="module-models_user-settings.html">models/user-settings</a></li><li><a href="module-routes.html">routes</a></li><li><a href="module-status.html">status</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 18:16:00 GMT-0500 (CDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-beta3</a> on Thu Apr 09 2015 22:09:57 GMT-0500 (CDT)
 </footer>
 
 <script> prettyPrint(); </script>
